--- src/main/java/com/example/demo/global/ApiResponse.java
@@ -1,29 +1,0 @@
-package com.example.demo.global;
-
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.annotation.JsonPropertyOrder;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-@JsonPropertyOrder({"isSuccess", "code", "message", "result"})
-public class ApiResponse<T> {
-
-    private final Boolean isSuccess;
-    private final String code;
-    private final String message;
-
-    @JsonInclude(JsonInclude.Include.NON_NULL)
-    private T result;
-
-    //성공시 응답
-    public static <T> ApiResponse<T> onSuccess(T result) {
-        return new ApiResponse<>(true, "COMMON200", "요청에 성공했습니다.", result);
-    }
-
-    //실패시 응답
-    public static <T> ApiResponse<T> onFailure(String code, String message, T data) {
-        return new ApiResponse<>(false, code, message, data);
-    }
-}
